+++ src/hooks/auth/useCompleteEmailSignup.ts
@@ -3,8 +3,8 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 
 import type { AuthResult } from '@/types/auth';
-import type { CompleteEmailSignupRequest } from '@/contracts/auth';
 import { authApi } from '@/api/auth/authApi';
+import type { CompleteEmailSignupRequest } from '@/contracts';
 
 export function useCompleteEmailSignup() {
   const queryClient = useQueryClient();
