+++ src/hooks/auth/useResetPassword.ts
@@ -2,8 +2,8 @@
 
 import { useMutation } from '@tanstack/react-query';
 
-import type { ResetPasswordRequest } from '@/contracts/auth';
 import { authApi } from '@/api/auth/authApi';
+import type { ResetPasswordRequest } from '@/contracts';
 
 export function useResetPassword() {
   return useMutation<void, Error, ResetPasswordRequest>({
