+++ src/hooks/auth/useUpdatePassword.ts
@@ -3,8 +3,8 @@
 import { useMutation } from '@tanstack/react-query';
 
 import type { AuthResult } from '@/types/auth';
-import type { UpdatePasswordRequest } from '@/contracts/auth';
 import { authApi } from '@/api/auth/authApi';
+import type { UpdatePasswordRequest } from '@/contracts';
 
 export function useUpdatePassword() {
   return useMutation<AuthResult, Error, UpdatePasswordRequest>({
