+++ src/hooks/auth/useVerifyEmailOtp.ts
@@ -2,11 +2,11 @@
 
 import { useMutation } from '@tanstack/react-query';
 
+import { authApi } from '@/api/auth/authApi';
 import type {
   VerifyEmailOtpRequest,
   VerifyEmailOtpResponse,
-} from '@/contracts/auth';
-import { authApi } from '@/api/auth/authApi';
+} from '@/contracts';
 
 export function useVerifyEmailOtp() {
   return useMutation<VerifyEmailOtpResponse, Error, VerifyEmailOtpRequest>({
